Other utility functions API

CePNEMAnalysis.gaussian_kernel — Function

gaussian_kernel(x::Real; sigma::Float64=1.0)

Evaluates a Gaussian kernel at position xwith standard deviationsigma`.

CePNEMAnalysis.convolve — Function

convolve(x::Vector{Float64}, k::Vector{Float64})

Convolves a vector x with a kernel vector k.

CePNEMAnalysis.correct_name — Function

correct_name(neuron_name)

Corrects the name of a neuron by removing the "0" substring from the name if it contains "DB0" or "VB0". This causes the neuron name to be compatible with the connectome files.

Arguments:

neuron_name::String: The name of the neuron to be corrected.

Returns:

neuron_name::String: The corrected name of the neuron.

CePNEMAnalysis.find_peaks — Function

find_peaks(neural_activity::Vector{Float64}, threshold::Float64)

Find the peaks in a vector of neural activity above a given threshold. This method is most useful for finding spikes in the activity of spiking neurons.

Arguments:

neural_activity::Vector{Float64}: A vector of neural activity.

threshold::Float64: The threshold above which to consider a value a peak.

Returns:

peaks::Vector{Int}: A vector of indices of the peaks in the neural activity.

peak_heights::Vector{Float64}: A vector of the heights of the peaks in the neural activity.